+++ package/Editor/Sources/Services/Packages/PackageTag.cs
@@ -1,0 +1,10 @@
+﻿namespace UnityEditor.PackageManager.UI
+{
+    public enum PackageTag
+    {
+        alpha,
+        beta,
+        recommended,
+        experimental
+    }
+}