+++ package/Editor/Sources/UI/PackageManagerWindow.cs
@@ -1,4 +1,4 @@
-﻿using UnityEngine;
+using UnityEngine;
 using UnityEngine.Experimental.UIElements;
 using UnityEditor.Experimental.UIElements;
 
@@ -8,6 +8,7 @@ namespace UnityEditor.PackageManager.UI
     {
         private const double targetVersionNumber = 2018.1;
 
+#if UNITY_2018_1_OR_NEWER
         // When object is created
         public void OnEnable()
         {
@@ -21,12 +22,20 @@ namespace UnityEditor.PackageManager.UI
             template.StretchToParentSize();
             this.GetRootVisualContainer().StretchToParentSize();
 
+            PackageSearchFilterTabs.visible = false;
+
             PackageList.OnSelected += OnPackageSelected;
+            PackageList.OnLoaded += OnPackagesLoaded;
         }
 
         private void OnPackageSelected(Package package)
         {
             PackageDetails.SetPackage(package, PackageSearchFilterTabs.CurrentFilter);
+        }
+
+        private void OnPackagesLoaded()
+        {
+            PackageSearchFilterTabs.visible = true;
         }
 
         private PackageList PackageList
@@ -47,24 +56,17 @@ namespace UnityEditor.PackageManager.UI
         [MenuItem("Project/Packages/Manage")]
         internal static void ShowPackageManagerWindow()
         {
-            // Check that we are using the right Unity version before we proceed.
-            // Eventually, we could launch different functionality here based on version.
-            var version = Application.unityVersion;
-
-            // version's format will be something like 2017.3.0b4.
-            // we need a numerical representation of the major.minor.
-            double versionNumber;
-
-            if (double.TryParse(version.Substring(0, version.LastIndexOf(".")), out versionNumber) && versionNumber < targetVersionNumber)
-            {
-                EditorUtility.DisplayDialog("Unsupported Unity Version", string.Format("The Package Manager requires Unity Version {0} or higher to operate.", targetVersionNumber), "Ok");
-                return;
-            }
-
             var window = GetWindow<PackageManagerWindow>(true, "Packages", true);
             window.minSize = new Vector2(850, 450);
             window.maxSize = window.minSize;
             window.Show();
         }
+#else
+        [MenuItem("Project/Packages/Manage")]
+        internal static void ShowPackageManagerWindow()
+        {
+            EditorUtility.DisplayDialog("Unsupported Unity Version", string.Format("The Package Manager requires Unity Version {0} or higher to operate.", targetVersionNumber), "Ok");
+        }
+#endif
     }
 }