+++ package/Tests/Editor/AssemblyInfo.cs
@@ -1,4 +1,3 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("com.unity.package-manager-captain.Editor")]
-[assembly: InternalsVisibleTo("com.unity.package-manager-ui.EditorTests")]