+++ package/Tests/Editor/Common/PackageSets.cs
@@ -79,7 +79,7 @@ namespace UnityEditor.PackageManager.UI.Tests
             {
                 version = string.Format("1.0.{0}", _count);
                 if (Random.NextDouble() > 0.5)
-                    version += "-preview";
+                    version += "-alpha";
             }
 
             var group = UpmBaseOperation.GroupName(type);
@@ -92,6 +92,7 @@ namespace UnityEditor.PackageManager.UI.Tests
                 State = PackageState.UpToDate,
                 Group = group,
                 Version = version,
+                IsRecommended = false,
                 IsCurrent = true,
                 IsLatest = false,
                 OriginType = type,